requirement.txt
streamlit
pandas
plotly
numpy
statsmodels
scikit-learn
pyyaml
joblib
